@@ -1,48 +1,41 @@
-[tool.poetry]
+[project]
 name = "fakts-next"
-version = "1.0.5"
+version = "1.1.0"
 description = "asynchronous configuration provider ( tailored to support dynamic client-server relations)"
-authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
-license = "CC BY-NC 3.0"
+authors = [{ name = "jhnnsrs", email = "jhnnsrs@gmail.com" }]
+requires-python = ">=3.11"
 readme = "README.md"
-packages = [{ include = "fakts_next" }]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-pydantic = ">2"
-PyYAML = ">=6"
-QtPy = "^2.0.1"
-koil = ">=1.0.0"
-aiohttp = { version = "^3.8.2", optional = true }
-certifi = { version = ">2021", optional = true }
-rich_click = { version = ">=0.3", optional = true }
-netifaces = { version = "^0.11.0", optional = true }
+license = "MIT"
+dependencies = [
+    "pydantic>2",
+    "koil>=1.0.3",
+    "certifi>=2025.4.26",
+    "aiohttp>=3.11.18",
+    "pyyaml>=6.0.2",
+]
 
 
-[tool.poetry.extras]
-remote = ["aiohttp", "certifi"]
-cli = ["rich_click","netifaces"]
+[project.scripts]
+fakts = "fakts.cli.main:cli"
 
+[tool.uv]
+dev-dependencies = [
+    "pyqt5>=5.15.11",
+    "pytest-asyncio>=0.26.0",
+    "pytest-cov>=6.1.1",
+    "pytest-qt>=4.4.0",
+    "pytest>=8.3.5",
+]
 
+[tool.hatch.build.targets.sdist]
+include = ["fakts_next"]
 
-[tool.poetry.group.dev.dependencies]
-httpx = "^0.23.0"
-autoflake = "^1.7.7"
-pytest = "^7.2.0"
-testcontainers = "^3.7.0"
-pytest-qt = "^4.2.0"
-pytest-asyncio = "^0.20.2"
-aiohttp = "^3.8.3"
-websockets = "^10.4"
-black = ">=22"
-pytest-cov = "^4.0.0"
-ruff = "^0.0.282"
-mypy = "^1.7.1"
-pydoc-markdown = "^4.8.2"
-types-toml = "^0.10.8.7"
-types-pyyaml = "^6.0.12.12"
-netifaces = "^0.11.0"
-rich-click = "^1.7.3"
+[tool.hatch.build.targets.wheel]
+include = ["fakts_next"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [tool.pytest.ini_options]
 qt_api = "pyqt5"
@@ -96,11 +89,3 @@ search_path = ["fakts"]
 [tool.pydoc-markdown.renderer]
 type = "docusaurus"
 docs_base_path = "website/docs"
-
-[tool.poetry.scripts]
-fakts = "fakts.cli.main:cli"
-
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
