@@ -1,7 +1,10 @@
 from collections.abc import MutableMapping
+from typing import Any, cast
 
 
-def update_nested(d: MutableMapping, u: MutableMapping) -> MutableMapping:
+def update_nested(
+    d: MutableMapping[str, Any], u: MutableMapping[str, Any]
+) -> MutableMapping[str, Any]:
     """Update a nested dictionary or similar mapping.
     This is a recursive function that will update the values in the dictionary
     *inplace*.
@@ -20,7 +23,7 @@ def update_nested(d: MutableMapping, u: MutableMapping) -> MutableMapping:
     """
     for k, v in u.items():
         if isinstance(v, MutableMapping):
-            d[k] = update_nested(d.get(k, {}), v)
+            d[k] = update_nested(d.get(k, {}), cast(MutableMapping[str, Any], v))
         else:
             d[k] = v
     return d
